First Commit: Define IQueryBuilder and implement it in MYSQLQueryBuilder
Define the and implement the error-management components(in Error.h and ErrorCategory.h).
Define the IMainclauseCreator, which is called when there is the need to create a query main-clasue.

Only SELECTClauseCretor(which handles the SELECT query-type of a MYSQL) implements the IMainClauseCreator and therefore, making MYSQLQueryCreator able to handle only "SELECT" query at the time of writing
Every implementation of the member-functions of MYSQLQueryBuilder, that accepts string arguements, starts with a comment inidacting that such arguments should be validated. Such validation mechanisms are yet to be added
All implementation of the member-functions of MYSQLQueryBuilder, that accepts string arguments, have comment sections which indicates that such arguments should be escaped(for MYSQL). Such escape mechanims are yet to be added